Fix: correct out_ierr type to real(dp) to match rel_rms_out argument

+++ vdart_solver_mod.f90
@@ -582,7 +582,7 @@ contains
     real(dp) :: FI0_AMP_prev
     character(len=32) :: tmpstr
     real(dp) :: mean_tq
-    integer :: out_ierr
+    real(dp) :: out_ierr
     character(len=128) :: csv_file
     integer :: csv_unit, openstat
